fix: size_t

+++ Backend/API/UserToken.hpp
@@ -20,7 +20,7 @@ std::string GenerateToken() {
       "abcdefghijklmnopqrstuvwxyz";
   std::string Result;
   Result.reserve(Length);
-  for (int i = 0; i < Length; ++i) {
+  for (size_t i = 0; i < Length; ++i) {
     Result += AlphaNum[rand() % (sizeof(AlphaNum) - 1)];
   }
   return Result;
